refactor: separate Study Block creation from its model
