+++ src/studyBlock.ts
@@ -1,7 +1,6 @@
 import { getFormattedParagraphs } from "./paragraphFormatting";
 import { GospelStudyPluginSettings } from "./models/GospelStudyPluginSettings";
 import { UrlParserResult } from "./models/UrlParserResult";
-import { requestUrl } from "obsidian";
 
 export class StudyBlock {
     private _paragraphs: string[] | undefined;
@@ -9,34 +8,12 @@ export class StudyBlock {
     private _tag: string | undefined;
 
     /**
-     * Creates a new StudyBlock instance using the given URL parser result and plugin settings.
-     * 
-     * @param urlParserResult - The result of parsing a URL.
-     * @param pluginSettings - The settings for the Gospel Study plugin.
-     * @returns A Promise that resolves to a StudyBlock instance.
-     */
-    public static async create(urlParserResult: UrlParserResult): Promise<StudyBlock> {
-        const urlToRequest = urlParserResult.url.toString();
-
-        const response = await requestUrl({
-            url: urlToRequest, method: "GET", headers: {
-                "cookie": "analytics_video_metadata_load=false"
-            }
-        });
-
-        const parser = new DOMParser();
-        const sourceDocument = parser.parseFromString(response.text, "text/html");
-        const studyBlock = new StudyBlock(urlParserResult, sourceDocument);
-
-        return studyBlock;
-    }
-
-    /**
      * Constructs a new instance of StudyBlock from the specified URL and source document.
      */
-    private constructor(
+    public constructor(
         private _urlParserResult: UrlParserResult,
-        private _sourceDocument: Document) { }
+        private _sourceDocument: Document
+    ) { }
 
     /**
      * Gets the string representation of the paragraph IDs excluding the "p" prefix.
